requirements.txt
dbt-core
dbt-duckdb
polars
requests
streamlit
